Change ElasticSearch default sort field (#737)
* The field data circuit breaker thrown an error in the TEST env because the memory used by the :id field in the field data cache exceeds the limit.

This commit replaces the default sort field :id by :_doc in Elasticsearch searches. The :id field is a text field that uses a large amount of memory in the field data cache due to its high cardinality. The :_doc is the most efficient sort and it doesn't use the field data cache.

* Use the latest version of clj-momo