Remove config for missing Riemann from demo (#774)
The ctia.properties in the demo docker container is configured to look for
Riemann, but no Riemann container is supplied.  This apparently causes CTIA
init to fail when trying to start the demo container.  Commenting out the
Riemann config in the `ctia/config/ctia.properties` file we supply to docker
seems to resolve the issue, allowing CTIA to start normally.

Resolves https://github.com/threatgrid/iroh/issues/2407