E3258: metric routes: topn, cardinality, histogram (#896)
> related https://github.com/threatgrid/iroh/issues/3258

This PR implements 3 metric endpoints for incident, sighting and judgement: `topn`, `cardinality` and`histogram`.
Each of these routes contains an `aggregate-on` parameter which accepted values are specific to each entity type. Then each metric type has it's own specific paremeters.
This PR also enables us to filter both search and metric requests with a date range using new `from` and/or `to` parameters. 
This PR extends the `IQueryStringSearchableStore` with an `aggregate` method, which is implemented in the ES store using [terms](https://www.elastic.co/guide/en/elasticsearch/reference/5.6/search-aggregations-bucket-terms-aggregation.html), [cardinality](https://www.elastic.co/guide/en/elasticsearch/reference/5.6/search-aggregations-metrics-cardinality-aggregation.html) and [date histogram](https://www.elastic.co/guide/en/elasticsearch/reference/5.6/search-aggregations-bucket-datehistogram-aggregation.html) aggregations in Elasticsearch. Each of these aggregations is performed against the results of a search query which is identical to `search` route behavior.
The `query-string-search` method is slightly refactored with to add a `date-range` option. In order to avoid inflating the number of search parameters, all search parameters are now embedded in a `search-query` map with following schema:
``` clojure
(s/defschema SearchQuery
  (st/optional-keys
   {:query-string s/Str
    :filter-map {s/Keyword s/Any}
    :date-range DateRange}))
```
There is many modifications that are only to enable `:query` to be optional.
The main modifications to focus on are the insertion of aggregation and from/to filters in following files: ` src/ctia/http/routes/crud.clj`, `src/ctia/http/routes/common.clj ` (search query build), `src/ctia/stores/es/crud.clj `. In test the whole logic to test metric routes is in this helper ns `test/ctia/test_helpers/aggregate.clj`, that generate examples of entities and call the 3 types of metric. The unit tests are performed at the lowest level. The route tests are mostly meant to check that parameters are properly passed to `handle-aggregate` in the ES store. `handle-aggregate` has its own tests that are responsible for checking the different intended behaviors.