strenghten string generation in metric tests (#907)
> Related #https://github.com/threatgrid/ctia/issues/905

Sometime the metric tests fail on labels. I did not manage to reproduce this issue.
This PR strengthen the label generator and improve test failing messages to help understanding if it happens again. I do not close the issue since I cannot confirm that it solves the problem, but it will help ease to improve this label generation if necessary.