Fix investigation schema (#910)
This PR Fixes Investigation Schemas. CTIM flanders schemas were internalized in CTIA, but the former schema were not removed and thus actually used instead of the new schema. 
Since the former schema was open, it did not cause any trouble on added fields, but this was the cause of missing fields in grapqhl.