use `id` field for sorting (#942)
> closes #https://github.com/threatgrid/iroh/issues/2870

the `id` field is of type `keyword` and stored as `doc_values` since the mapping migration. This PR makes the migration sort on `id` instead of `_uid` to definitively get rid of fielddata (sorting on `_uid` is performed with fielddata on elaticsearch 5.6, which thus consume a lot of memory due to the cardinality of this field.