move migration params in properties (#943)
> Related #threatgrid/iroh#3072

This PR moves the migration parameters in properties. 
These parameters were passed through the command line argument. For the next migration we need to add more arguments, in particular arguments for the target cluster and it will excessively increase the command line. Thus we decided to move these parameters in configuration. 
This is in particular the first step toward decorrelating the source cluster properties and target cluster properties for the ES7 migration.