Fix pagination (#993)
> Close #threatgrid/iroh/issues/4192

The inconsistent results in feeds are due to the `all-pages` function that performs a `search_after` based pagination using the default `sort_by` parameter. However this default value is base on `_doc` field which is unique per shard but not per index! Indeed this is a problem that can occur for any `search_after` pagination that will use the default `sort_by` value. 
This PR changes the default `sort_by` value to `_doc,id` that will first sort by `_doc` to ensure the order then by `id` for disambiguation. It also generalizes tests with multiple shards to better reflect production properties.