fix ductile protocol parameter (#1008)
> Related #threatgrid/iroh/issues/4216
> Related #threatgrid/iroh/issues/3072

This PR fixes what I broke :-).
Ductile uses a parameter `protocol` instead of `transport` to precise the uri protocol in the ESConn.
I was not used in CTIA tests neither in any of our deployed instances that use the default value (`http`). Thus this bug did not appear when I introduced ductile.