make configurable auto mappings/settings update (#1)
Updating an existing field mapping with `include_in_all` does not make the mapping update fail like any other update of existing field mapping. We need to disable it until ES7 migration, and have a simple control on this.
This PR enable disable mapping/setting update directly from the properties.
The default value is false for a better control on the indices life cycle.