Es7 init (#1013)
> Close #https://github.com/threatgrid/iroh/issues/4220
> Related #https://github.com/threatgrid/iroh/issues/4217
> Related #https://github.com/threatgrid/iroh/issues/4288

This PR enables CTIA to run with either Elasticsearch 5 or 7. it impacts most store operations including:
- mapping
- index initialization
- CRUD
- bulk
switching between ES 5 and ES 7 is possible throw the configuration by simply setting the parameter `ctia.store.default.es.version`. It is also configurable per store like other ES parameters and indeed we could migrate iteratively stores and have stores on ES5 and some stores on ES7 while performing the migration.

⚠️ I must disable the mapping update. I prepared the switch to ES7 which requires mapping changes of existing fields. Once this PR will be merged it will be not possible to add new fields mapping through CTIA task. If anyone needs to introduce new fields before the migration, I can update mappings manually if you tell me a nice joke.
Updating an existing field mapping with `include_in_all` does not make the mapping update fail like any other update of existing field mapping. We need to disable it until ES7 migration, and have a simple control on this. This PR includes the ability to disable mapping/setting update directly from the properties. The default value is false for a better control on the indices life cycle.


For compatibility reason it impacts the migration task and the removal of `include_in_all` which are also mostly handled here (with some few details that will be handled in coming PR dedicated to this).
The compatibility between the 2 versions are mostly delegated to our new ES library, ductile. Depending on the queried ES version, which is now part of the connection fields, ductile builds the proper uri/parameters. I kept the corresponding PR opened: https://github.com/threatgrid/ductile/pull/5.

Since I tried to delegate as far as possible the compatibility to ductile, most of the changes are in the tests to enable testing both ES 5 and ES 7. An helper macro slightly derived from the one used in ductile is used for testing both version. The `Hide whitespace changes` should protect your eyes :-).

As a reminder the migration process is as follow, for each ENV:
- let CTIA runs on current Elasticsearch cluster and indices
- start migration task (4 billion documents on NAM private CTIA, should take 1 week)
- when migration task is complete restart on new cluster / indices
As mentioned this process could be decompose for each ENV / store.