configurable default sort (#1163)
This PR introduces a configurable default sort for searching.
I could not reproduce the issue locally, but for some reason the sorting on `_doc` is inconsistent even if we also sort by id as a second sort element (which [previously fixed inconsistency for feeds](https://github.com/threatgrid/ctia/pull/993)).
My intuition is that it happens when different replicas or primary shards are hit. During the migration, we turn off the replication for performance reason and then turn on the replication at the end. It's very likely that a parallelization during the replication do not ensure that the `_doc` values are preserved across replicas. I could not replicate this behavior.