Reduce noise in logs when running tests (#1192)
> Related https://github.com/threatgrid/ctia/pull/1189
> Related https://github.com/threatgrid/ctia/pull/1190#issuecomment-1025980043

Adds a new configuration file for logback that change the configuration during the tests. By default, it disables logs in the root logger and only allows logger that start with `ctia`.

During its initialization, logback tries to find a file logback-test.xml in the classpath. If no such file is found, it checks for the file logback.xml in the classpath.

This PR adds a `logback-test.xml` file which is only available in the test resource folder. It will not be deployed by `uberjar` or used with the command `lein run -m ctia.main`.

<a name="qa">[§](#qa)</a> QA
============================

No QA is needed.

<a name="release-notes">[§](#release-notes)</a> Release Notes
=============================================================

```
intern: Disable logs during tests
```