Add id into sort set
This should cover the case when two relationships was created for the
same entity (corelated via :source_ref or :target_ref) at the same time
with milliseconds precision. Without id in the sort set :search_after
might not work as expected because it might just jump to the next
record.