Fix bundle export (#1291)
> Related advthreat/iroh#4958 #1285 

Performance improvements PR contains a bug where identity is passed as a record instead of a map into a function to fetch related nodes. Unfortunately, implementation silently delists records as inaccessible, making responses almost empty. This PR fixes that by normalising arglist for the bundle-export function.

<a name="qa">[§](#qa)</a> QA
============================

QA instructions are in #1285

<a name="release-notes">[§](#release-notes)</a> Release Notes
=============================================================

<a name="squashed-commits">[§](#squashed-commits)</a> Squashed Commits
======================================================================

Attempt to fix the problem when some entities do not load as a part
of the export bundle.