threatgrid/ctim/#381 Migrate actor 1.2.0 (#1323)
Related threatgrid/ctim#381

Update actor to reflect new CTIM model, in which actor_types is modified to be both optional and an array. Add migration code.

I ran the migration locally following the directions in the migration.md doc and it worked fine. I also confirmed that I was able to search by actor_types afterward.